refactor: merge FSEntryFetcher with FSEntryController
